Memory

+++ src/core/StringBuilder.hpp
@@ -125,7 +125,7 @@ private:
 	{
 		if (_capacity > capacity) return;
 
-		_capacity = Math::Max(8U, _capacity);
+		_capacity = Math::Max((size_t)8, _capacity);
 		while (_capacity < capacity) {
 			_capacity *= 2;
 		}
